Enum clog::error::Error

pub enum Error {
    ConfigParseErr,
    ConfigFormatErr,
    CurrentDirErr,
    TomlReadErr,
    LinkStyleErr,
    SemVerErr,
    CreateFileErr,
    WriteErr,
    IoErr,
    UnknownErr,
}

An enum for describing and handling various errors encountered while dealing with clog building, or writing of changelogs.

Variants

ConfigParseErr

Generated while parsing config files such .clog.toml

ConfigFormatErr

Generated if the config file is not in TOML format

CurrentDirErr

Generated if the OS cannot determine the current directory

TomlReadErr

Generated when unable to read the TOML config file (perhaps due to permissions, etc.)

LinkStyleErr

Generated when an unrecognized link-style value is used

SemVerErr

Generated when a string cannot be parsed to a SemVer value

CreateFileErr

Generated when there are errors creating the output file or stream

WriteErr

Generated when there are errors writing to the changelog output file

IoErr

Generic catch all I/O related error

UnknownErr

Unknown, but fatal error (a catch all)

Methods

impl Error

fn is_fatal(&self) -> bool

Return whether this was a fatal error or not.

fn exit(&self) -> !

Print this error and immediately exit the program.

If the error is non-fatal then the error is printed to stdout and the exit status will be 0. Otherwise, when the error is fatal, the error is printed to stderr and the exit status will be 1.