Module clog::fmt

Structs

JsonWriter
Wraps a std::io::Write object to write clog output in a JSON format

MarkdownWriter
Wraps a std::io::Write object to write clog output in a Markdown format

Enums

ChangelogFormat

Traits

FormatWriter
A trait that allows writing the results of a clog run which can then be written in an arbitrary format. The single required function write_changelog() accepts a clog::SectionMap which can be thought of similiar to a clog "AST" of sorts.

Type Definitions

WriterResult
Convienience type for returning results of writing a changelog with a Clog struct